backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
reportlab==4.0.8
supabase==2.4.0
python-dotenv==1.0.1
python-dotenv==1.0.1
